# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Financials API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/investment-returns": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "occurredOn",
                  "amount",
                  "accountId"
                ],
                "properties": {
                  "occurredOn": {
                    "description": "Calendar date, YYYY-MM-DD",
                    "type": "string"
                  },
                  "amount": {
                    "description": "Signed decimal string, never zero, up to 12 integer digits and 2 decimals",
                    "type": "string"
                  },
                  "accountId": {
                    "description": "Any account of the user, active or not",
                    "type": "string"
                  },
                  "notes": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "accountId",
                    "accountNickname",
                    "occurredOn",
                    "amount",
                    "notes"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "accountId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "accountNickname": {
                      "type": "string"
                    },
                    "occurredOn": {
                      "description": "Calendar date, YYYY-MM-DD",
                      "type": "string"
                    },
                    "amount": {
                      "description": "Signed decimal string with 2 decimals, never zero, e.g. \"50.00\" or \"-20.00\"",
                      "type": "string"
                    },
                    "notes": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "lastDate"
                  ],
                  "properties": {
                    "items": {
                      "description": "Newest date first",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "accountId",
                          "accountNickname",
                          "occurredOn",
                          "amount",
                          "notes"
                        ],
                        "properties": {
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "accountId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "accountNickname": {
                            "type": "string"
                          },
                          "occurredOn": {
                            "description": "Calendar date, YYYY-MM-DD",
                            "type": "string"
                          },
                          "amount": {
                            "description": "Signed decimal string with 2 decimals, never zero, e.g. \"50.00\" or \"-20.00\"",
                            "type": "string"
                          },
                          "notes": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "lastDate": {
                      "description": "Date of the newest return (YYYY-MM-DD), null when there are none",
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/investment-returns/{id}": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "occurredOn": {
                    "description": "Calendar date, YYYY-MM-DD",
                    "type": "string"
                  },
                  "amount": {
                    "description": "Signed decimal string, never zero, up to 12 integer digits and 2 decimals",
                    "type": "string"
                  },
                  "accountId": {
                    "description": "Any account of the user, active or not",
                    "type": "string"
                  },
                  "notes": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "accountId",
                    "accountNickname",
                    "occurredOn",
                    "amount",
                    "notes"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "accountId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "accountNickname": {
                      "type": "string"
                    },
                    "occurredOn": {
                      "description": "Calendar date, YYYY-MM-DD",
                      "type": "string"
                    },
                    "amount": {
                      "description": "Signed decimal string with 2 decimals, never zero, e.g. \"50.00\" or \"-20.00\"",
                      "type": "string"
                    },
                    "notes": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}